refactor: migrate to TypeScript
This change migrates the project to TypeScript.  Beyond providing types now, as well as an ESM option, the aim was to not change the API in a breaking way.

In doing this migration, I actually found at least one bug (TypeScript doing its job!).  In `validate`, `getSpecMap` was called, and immediately after there was an if condition to check if the result that came back was `false`, which would indicate the requested spec name was bad.  However the if condition checked if `specName` was false, rather than the result of the `getSpecMap` function.  That would have allowed people to pass in nonsense specNames and throwing an error later when attempting to access `map`.  I've fixed that here.

I did put a TODO in though to consider deprecating the one named PJV export, in favor of directly exporting the functions and constants.  This would make the API more tree-shakable and (in my opinion) more straightforward.  But that should be a separate change, if implemented.

The other quality of life stuff I adjusted:
- Added TsDoc in more places
- Changed concatenated strings to template strings