feat(validate): expand errors to include field name
This change adds changes the `errors` array that's returned from `validate` to be an array of objects that includes `field` and `message`, so that each error message is associated with which field the error applies to.

Question: is there value in doing similar for `warnings` and `recommendations`?  I'm thinking not, since the presence of those won't make the overall result invalid (whereas `errors` does result in an invalid result), but i'm certainly open to applying the same to those for consistency.

BREAKING CHANGE