feat(validateBin): add new validateBin function (#243)
Recreation of #230 with the breaking change notice removed. 

## PR Checklist

- [x] Addresses an existing open issue: fixes #223
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateBin` function, which checks that the
value passed in is valid for assigning to the "bin" property of a
`package.json`

Criteria: it's either a non-empty string or a Record<string, string>
with command name to command path.

The `validate` function's validation of `bin` is now stricter than
before.

Closes #223

Co-authored-by: @michaelfaith

Co-authored-by: michael faith <michaelfaith@users.noreply.github.com>