feat(validate): expand errors to include field name (#242)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #51
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds changes the `errors` array that's returned from
`validate` to be an array of objects that includes `field` and
`message`, so that each error message is associated with which field the
error applies to.

`errors` is now of type `ValidationError[]` where `ValidationError` is
```ts
interface ValidationError {
  field: string;
  message: string;
}
```

Question: is there value in doing similar for `warnings` and
`recommendations`? I'm thinking not, since the presence of those won't
make the overall result invalid (whereas `errors` does result in an
invalid result), but i'm certainly open to applying the same to those
for consistency.