build: bundle build files
This change is focused on optimizing our package, and to address a known bug with how we were handling support for both cjs and esm.

With this change, we're now bundling the two main entry points:
- the root of our js/ts api
- the cli bin

As part of multi-entry point best practices, I changed the import of `validate` in the second entry-point to references the package, rather than relative path, so we weren't doubling up on the bundles.

Package size (compressed) before:
Package size (compressed) after: 15kb