feat(validateType): add function for validating type
This change adds a new `validateType` function that validates the value of the "type" field of a package.json.  If returns a list of error messages if a violation is found.

The change also adds `type` as a recommended field for the `'npm'` spec in `validate`.  If `type` is absent from the package data passed into `validate`, it'll add a message to `recommendations`  in the return object.