feat(validateType): add function for validating type  (#284)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #68 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateType` function that validates the value
of the "type" field of a `package.json`. It returns a list of error
messages if a violation is found.

The change also adds `type` as a recommended field for the `'npm'` spec
when running `validate`. If `type` is absent from the package data
passed into `validate`, it'll add a message to `recommendations` in the
return object.