feat(validateScripts): add function for validating `scripts`
This change adds a new `validateScripts` function that validates the value of the "scripts" field of a package.json.  If returns a list of error messages if a violation is found.  Validation of scripts in the monolithic `validate` function has been switched over to use this function as well.  That means the criteria for `scripts` when running `validate` is stricter than it was before.

The new function validates scripts against the following criteria:
- the property is an object
- its keys are non-empty strings
- its values are all non-empty strings