feat(validateBundleDependencies): add function for validating `bundleDependencies`
This change adds a new `validateBundleDependencies` function that validates the value of the "bundleDependencies" field of a package.json. If returns a list of error messages if a violation is found. Validation of bundleDependencies in the monolithic validate function has been switched over to use this function as well. That means the criteria for bundleDependencies when running validate is stricter than it was before.

The new function validates scripts against the following criteria:

- the property is either an array or a boolean
- if it's an array, all items should be strings

Previously, it only validated that the type was an array.