feat: deprecate support for legacy commonjs specifications
This change officially deprecates support for the `commonjs_1.0` and `commonjs_1.1` specs in both the js api and cli.  The NPM spec will be the only supported spec.

Per our deprecation policy, following this release of this deprecation, the api is subject to be removed after 6 months.

Please begin using the `validate(data, options?)`  form of the JS API.