feat(validateBundleDependencies): add function for validating `bundleDependencies` (#308)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #232 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateBundleDependencies` function that
validates the value of the "bundleDependencies" field of a package.json.
If returns a list of error messages if a violation is found. Validation
of bundleDependencies in the monolithic validate function has been
switched over to use this function as well. That means the criteria for
bundleDependencies when running validate is stricter than it was before.

The new function validates scripts against the following criteria:

- the property is either an array or a boolean
- if it's an array, all items should be strings

Previously, it only validated that the type was an array.