feat(validateConfig): add function for validating `config`
This change adds a new `validateConfig` function that validates the value of the "config" field of a package.json. If returns a list of error messages if a violation is found. Validation of `config` in the monolithic validate function has been switched over to use this function as well. Though, the criteria is the same.

The new function simply validates that the value is an object.