feat(validateLicense): add function for validating `license`
This change adds a new `validateLicense` function that validates the value of the "license" field of a package.json. If returns a list of error messages if a violation is found. Validation of `license` in the monolithic validate function has been switched over to use this function as well. That means the criteria for license, when running validate, is stricter than it was before.

The new function uses `validate-npm-package-license` to validate the value (after first checking that it's a non-empty string).  Previously it was just validating that the value was a string.