feat(validateLicense): add function for validating `license` (#346)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #344 #341 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateLicense` function that validates the
value of the "license" field of a package.json. If returns a list of
error messages if a violation is found. Validation of `license` in the
monolithic validate function has been switched over to use this function
as well. That means the criteria for license, when running validate, is
stricter than it was before.

The new function uses `validate-npm-package-license` to validate the
value (after first checking that it's a non-empty string). Previously it
was just validating that the value was a string.