feat(validateVersion): add function for validating `version`
This change adds a new `validateVersion` function that validates the value of the "version" field of a package.json. It returns a list of error messages if a violation is found. Validation of `version` in the monolithic validate function has been switched over to use this function as well. That means the criteria for version, when running validate, is possibly stricter.

The new function uses `semver` to validate the value (after first checking that it's a non-empty string).  Previously it was using a bespoke regex pattern to validate.