feat(validateCpu): add function for validating `cpu`
This change adds a new `validateCpu` function that validates the value of the "cpu" field of a package.json. It returns a list of error messages if a violation is found. Validation of `cpu` in the monolithic validate function has been switched over to use this function as well. That means the criteria for `cpu`, when running `validate`, is stricter.

The new function validates that the value is an array of non-empty strings.  Previously it only checked that the value was an array.