feat(validateCpu): add function for validating `cpu` (#361)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #310
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateCpu` function that validates the value
of the "cpu" field of a package.json. It returns a list of error
messages if a violation is found. Validation of `cpu` in the monolithic
validate function has been switched over to use this function as well.
That means the criteria for `cpu`, when running `validate`, is stricter.

The new function validates that the value is an array of non-empty
strings. Previously it only checked that the value was an array.