feat(validateDependencies): add function for validating `dependencies`
This change adds a new `validateDependencies` function that validates the value of the "dependencies" field of a package.json. It returns a list of error messages if a violation is found. Validation of `dependencies`, `optionalDependencies`, `devDependencies`, and `peerDependencies` in the monolithic validate function have been switched over to use this function as well. That means the criteria for these properties, when running `validate`, is stricter.

In addition to the previous validations, it uses `semver` validate version ranges, and does more to validate the shape of the object.