feat(validateDependencies): add function for validating `dependencies` (#365)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #312, #342
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateDependencies` function that validates
the value of the "dependencies" field of a package.json. It returns a
list of error messages if a violation is found. Validation of
`dependencies`, `optionalDependencies`, `devDependencies`, and
`peerDependencies` in the monolithic validate function have been
switched over to use this function as well. That means the criteria for
these properties, when running `validate`, is stricter.

In addition to the previous validations, it uses `semver` validate
version ranges, and does more to validate the shape of the object.