feat(validateDescription): add function for validating `description`
This change adds a new `validateDescription` function that validates the value of the "description" field of a package.json. It returns a list of error messages if a violation is found. Validation of `description` in the monolithic validate function has been switched over to use this function as well. That means the criteria for `description`, when running `validate`, is stricter.

The new function validates that the value is a non-empty string.