feat(validateDirectories): add function for validating `directories`
This change adds a new `validateDirectories` function that validates the value of the "directories" field of a package.json. It returns a list of error messages if a violation is found. Validation of `directories` in the monolithic validate function has been switched over to use this function as well. That means the criteria for `directories`, when running `validate`, is stricter.

The new function validates that the value is a record of non-empty string keys to non-empty string values.