build: change relative import extensions to ts
This change moves all of our relative path imports to use `.ts` extensions instead of `.js`.  In order to do this, and avoid a bug in `tsup`, we're moving to use `tsdown` for bundling.  There should be no user-facing changes from this update.

I also changed the ci workflow to default node version to the latest lts version.