feat(validateExports): add function for validating `exports`
This change adds a new `validateExports` function that validates the value of the "exports" field of a package.json. It returns a list of error messages if a violation is found. The monolithic `validate` function, wasn't checking `exports` previously.  So, this is a net-new check for that function.

The new function validates that the value is either a string or an object with values that are either an entry path string or another exports condition object.