feat(validateExports): add function for validating `exports` (#395)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #362
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateExports` function that validates the
value of the "exports" field of a package.json. It returns a list of
error messages if a violation is found. The monolithic `validate`
function, wasn't checking `exports` previously. So, this is a net-new
check for that function.

The new function validates that the value is either a string or an
object with values that are either an entry path string or another
exports condition object.