refactor: move the Result API from function-based to class-based (#507)
## PR Checklist

- [x] Addresses an existing open issue: fixes #503
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change overhauls the new Result api to be class-based. The core of
the new implementation is the `Result` class and `ChildResult` subclass.
Everything that was previously possible with the functional api (and a
bit more) is now available from methods on these two classes.