feat(validateHomepage): add function for validating `files`
This change adds a new `validateHomepage` function that validates the value of the "homepage" field of a package.json. It returns a Result object with any issues detected.

The new function is effectively the same as what the monolithic `validate` function was testing: that the value is a string that matches a url format.