feat(validateMan): add function for validating `main`
This change adds a new `validateMan` function that validates the value of the "man" field of a package.json. It returns a Result object with any issues detected.

The new function is stricter than the existing `validate` function.  It only checked that the value was a string or array.  This function checks that but also validates that the string(s) end in a number (and optionally .gz).