feat(validateKeywords): add function for validating `keywords` (#533)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #373
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateKeywords` function that validates the
value of the "keywords" field of a package.json. It returns a Result
object with any issues detected.

The new function is slightly stricter, than what the monolithic check
for keywords was doing. It was only checking that the property had an
array value. This is doing that but also checking that all items are
non-empty strings.