feat(validateOs): add function for validating `os`
This change adds a new `validateOs` function that validates the value of the "os" field of a package.json. It returns a Result object with any issues detected.

The new function is stricter than the existing `validate` function.  It only checked that the value an array.  This function checks that but also validates that the items in the array are all strings from the list of possible node platforms.