feat(validatePublishConfig): add function for validating `os`
This change adds a new `validatePublishConfig` function that validates the value of the "publishConfig" field of a package.json. It returns a Result object with any issues detected.

The new function is much stricter than what the monolith `validate` function was checking.  It was only checking if the value was an object.  Now we're doing that, but also checking the values of specific properties (when present) that are known to be supported by npm, pnpm, and yarn.  I didn't include every single property that these package managers support, but picked the ones that seemed like they'd be used the most.  If we need to add more, it's fairly easy to do.