feat(validateEngines): add function for validating `engines`
This change adds a new `validateEngines` function that validates the value of the "engines" field of a package.json. It returns a Result object with any issues detected.

The new function is a bit stricter than what the monolith `validate` function was checking.  It was only checking if the value was an object.  Now we're doing that, but also checking that the values and keys are all non-empty strings.