feat(validateName): add function for validating `name`
This change adds a new `validateName` function that validates the value of the "name" field of a package.json. It returns a Result object with any issues detected.

The new function is much more accurate than the check that the monolithic `validate` function used to be doing.  It used its own bespoke regex pattern to detect a bad name.  The new function is using `validate-npm-package-name` to do it, which is the same package `npm` uses.