feat(validateName): add function for validating `name` (#551)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #536
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateName` function that validates the value
of the "name" field of a package.json. It returns a Result object with
any issues detected.

The new function is much more accurate than the check that the
monolithic `validate` function used to be doing. It used its own bespoke
regex pattern to detect a bad name. The new function is using
`validate-npm-package-name` to do it, which is the same package `npm`
uses.