feat(validateOs): add function for validating `os` (#544)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #376
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateOs` function that validates the value of
the "os" field of a package.json. It returns a Result object with any
issues detected.

The new function is stricter than the existing `validate` function. It
only checked that the value an array. This function checks that but also
validates that the items in the array are all strings from the list of
possible node platforms.