feat(validatePublishConfig): add function for validating `publishConfig` (#545)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #378
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validatePublishConfig` function that validates
the value of the "publishConfig" field of a package.json. It returns a
Result object with any issues detected.

The new function is much stricter than what the monolith `validate`
function was checking. It was only checking if the value was an object.
Now we're doing that, but also checking the values of specific
properties (when present) that are known to be supported by npm, pnpm,
and yarn. I didn't include every single property that these package
managers support, but picked the ones that seemed like they'd be used
the most. If we need to add more, it's fairly easy to do.