feat(validateSideEffects): add function for validating `sideEffects`
This change adds a new `validateSideEffects` function that validates the value of the "sideEffects" field of a package.json. It returns a Result object with any issues detected.

This is an entirely new validation for this package.  It checks that the value is either a `boolean` or an `Array` of non-empty strings.