feat(validateSideEffects): add function for validating `sideEffects` (#571)
<!-- 👋 Hi, thanks for sending a PR to package-json-validator! 💖
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #570
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/package-json-validator/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/package-json-validator/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

This change adds a new `validateSideEffects` function that validates the
value of the "sideEffects" field of a package.json. It returns a Result
object with any issues detected.

This is an entirely new validation for this package. It checks that the
value is either a `boolean` or an `Array` of non-empty strings.

---------

Co-authored-by: Josh Goldberg ✨ <git@joshuakgoldberg.com>