build: switch from markdownlint to @eslnt/markdown
This change takes inspiration from https://github.com/JoshuaKGoldberg/create-typescript-app/pull/2298 and removes markdownlint in favor of @eslint/markdown.  Switching to the new plugin exposed that our Changelog headings aligned.  So this also switches our release-it config to use the `conventionalcommits` preset, instead of the `angular` one, and updates the `CHANGELOG.md` to conform.