ci: move releases from release-it to release-please
This change updates our `release` workflow to use `release-please` instead of `release-it`.  This will create a new Release PR for adding changelog entries and bumping package version, as new changes are merged into `main`.  At the point where you want to do the release, a maintainer would just merge the release PR.  That triggers the build and publish steps to run.